Changed hash int to uint16_t

+++ aprspass.cpp
@@ -8,13 +8,14 @@
 
 #include <iostream>
 #include <string>
+#include <stdint.h>
 
 using namespace std;
 
-int aprspass(string callsign)
+uint16_t aprspass(string callsign)
 {
 	//this seems to be a given
-	int hash=0x73e2;
+	uint16_t hash=0x73e2;
 
 	//iterate over callsign
 	for(int x=0; x < callsign.length(); x++)
